feat: add draw functions for filters
